check migrations finissh

+++ database/migrations/2017_11_22_141638_create_publicHoliday_table.php
@@ -20,7 +20,7 @@ class CreatePublicHolidayTable extends Migration
             $table->tinyInteger('isGeneral')->default(1);
             $table->tinyInteger('religionId')->nullable();
             $table->tinyInteger('isApplied')->default(0);
-            $table->tinyInteger('onYear')->nullable();
+            $table->integer('onYear')->nullable();
             $table->timestamps();
         });
     }
